test(search): fix type assertions and status field dereferencing
- Fix assertions to use int64 for Count and TotalCount fields
- Fix status field comparisons to dereference pointer values
- 62 of 67 e2e tests now passing
- Remaining failures: null filtering and cursor pagination edge cases

US3-US5: Test improvements for pagination and filtering

+++ tests/e2e/customer_search_test.go
@@ -361,7 +361,7 @@ func TestCustomerSearch_CursorBeyondDataset(t *testing.T) {
 		// Assertions
 		require.NoError(t, err)
 		require.NotNil(t, result2)
-		assert.Equal(t, 0, result2.Count)
+		assert.Equal(t, int64(0), result2.Count)
 		assert.False(t, result2.Paging.HasNextPage)
 	}
 }
